requirements.txt
kafka-python
faker
psycopg2-binary
